# apigateway/docs/swagger.yaml
definitions:
...
  response.ApiResponsePaginationProduct:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ProductResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationProductDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.ProductResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponseProduct:
    properties:
      data:
        $ref: '#/definitions/response.ProductResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseProductAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseProductDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseProductDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.ProductResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
  response.PaginationMeta:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.ProductResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  response.ProductResponseDeleteAt:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/product:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationProduct'
        "500":
          description: Failed to retrieve product data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all products
      tags:
      - Product
  /api/product/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product data
          schema:
            $ref: '#/definitions/response.ApiResponseProduct'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve product data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find product by ID
      tags:
      - Product
  /api/product/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active products
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationProductDeleteAt'
        "500":
          description: Failed to retrieve product data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active products
      tags:
      - Product
  /api/product/create:
    post:
      consumes:
      - multipart/form-data
      description: Create a new product with the provided details and an image file
      parameters:
      - description: Merchant ID
        in: formData
        name: merchant_id
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: string
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product description
        in: formData
        name: description
        required: true
        type: string
      - description: Product price
        in: formData
        name: price
        required: true
        type: number
      - description: Product count in stock
        in: formData
        name: count_in_stock
        required: true
        type: integer
      - description: Product brand
        in: formData
        name: brand
        required: true
        type: string
      - description: Product weight
        in: formData
        name: weight
        required: true
        type: number
      - description: Product rating
        in: formData
        name: rating
        required: true
        type: number
      - description: Product slug
        in: formData
        name: slug_product
        required: true
        type: string
      - description: Product image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created product
          schema:
            $ref: '#/definitions/response.ApiResponseProduct'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create product
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new product
      tags:
      - Product
  /api/product/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a product record by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted product record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseProductDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete product:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a product
      tags:
      - Product
  /api/product/delete/all:
    post:
      consumes:
      - application/json
      description: Permanently delete all product records.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted all product records permanently
          schema:
            $ref: '#/definitions/response.ApiResponseProductAll'
        "500":
          description: Failed to delete all products
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete all products
      tags:
      - Product
  /api/product/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed product record by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored product
          schema:
            $ref: '#/definitions/response.ApiResponseProductDeleteAt'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore product
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed product
      tags:
      - Product
  /api/product/restore/all:
    post:
      consumes:
      - application/json
      description: Restore all trashed product records.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored all products
          schema:
            $ref: '#/definitions/response.ApiResponseProductAll'
        "500":
          description: Failed to restore all products
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore all trashed products
      tags:
      - Product
  /api/product/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed products
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationProductDeleteAt'
        "500":
          description: Failed to retrieve product data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed products
      tags:
      - Product
  /api/product/trashed/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a trashed product record by its ID.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed product
          schema:
            $ref: '#/definitions/response.ApiResponseProductDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed product
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed product
      tags:
      - Product
  /api/product/update/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Update an existing product record with the provided details and
        an optional image file
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Merchant ID
        in: formData
        name: merchant_id
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: string
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product description
        in: formData
        name: description
        required: true
        type: string
      - description: Product price
        in: formData
        name: price
        required: true
        type: number
      - description: Product count in stock
        in: formData
        name: count_in_stock
        required: true
        type: integer
      - description: Product brand
        in: formData
        name: brand
        required: true
        type: string
      - description: Product weight
        in: formData
        name: weight
        required: true
        type: number
      - description: Product rating
        in: formData
        name: rating
        required: true
        type: number
      - description: Product slug
        in: formData
        name: slug_product
        required: true
        type: string
      - description: Product image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated product
          schema:
            $ref: '#/definitions/response.ApiResponseProduct'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update product
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing product
      tags:
      - Product
...
swagger: "2.0"